Update info on "Jenkins community account" (#1609)
During the hosting process, I have helped a few people with upload permissions issues. They were not aware that the "Jenkins community account" is the same as the one they might login to Jira with. I just added a short blurb about this to the description to hopefully reduce this confusion.