Add darinpope and markewaite as embeddable build status maintainers
https://docs.google.com/document/d/11Nr8QpqYgBiZjORplL_3Zkwys2qK1vEvK-NYyYa4rzg/edit?disco=AAAAYXxgRHI
provides more context for the adoption request.  We believe it may be
cheaper and easier to adopt this plugin than to remove the plugin from
ci.jenkins.io and update the many repositories that refer to it.

Help desk https://github.com/jenkins-infra/helpdesk/issues/3013 suggests
that we should consider removing the plugin because it is not actively
maintained.

Help desk https://github.com/jenkins-infra/helpdesk/issues/3044 proposes
steps to replace the use of this plugin with something hosted elsewhere
on Jenkins infrastructure.

One of the existing maintainers needs to approve the adoption request.
Existing maintainers are:

* @thomas-dee
* @mgedmin
* @christiangalsterer
* @jglick