Remove Mark, Jean-Marc, and Bruno from several plugins
Plugins were adopted for the "Contributing to Open Source" workshop at
DevOps World 2022.  The workshop was cancelled when DevOps World was
postponed.  Changes from Mark, Jean-Marc, and Bruno have been merged to
many of these plugins.  Returning them to the pool of plugins that are
up for adoption.

Confirmed in a private meeting with @jmMeessen and @gounthar that
they approve.  If desired, merge can wait until they approve on this
pull request.