Add markewaite as an xshell plugin maintainer
I use this plugin to validate some bug fixes related to freestyle jobs
in my instance.

The Jenkins baseline that it requires is old enough that it
causes an implicit dependency on the windows-slaves-plugin.
The windows-slaves-plugin plugin will be deprecated in May 2023 when
Microsoft bans the DCOM technique required by the windows-slave-plugin.
I don't want deprecated plugins in my installation, so I need to increase
the minimum Jenkins version of this plugin so that it no longer has an
implicit dependency on windows-slaves-plugin.

The plugin site describes implied dependencies as follows:

Features are sometimes detached (or split off) from Jenkins core and
moved into a plugin. Many plugins, like Subversion or JUnit, started as
features of Jenkins core.

Plugins that depend on a Jenkins core version before such a plugin was
detached from core may or may not actually use any of its features. To
ensure that plugins don't break whenever functionality they depend on
is detached from Jenkins core, it is considered to have a dependency on
the detached plugin if it declares a dependency on a version of Jenkins
core before the split. Since that dependency to the detached plugin is
not explicitly specified, it is implied.

Plugins that don't regularly update which Jenkins core version they
depend on will accumulate implied dependencies over time.