Prepare to deprecate unleash Maven plugin
https://github.com/shillner/unleash-maven-plugin/ has been archived.
That is the unleash plugin for Apache Maven.  The Jenkins unleash
maven plugin depends on the unleash plugin for Apache Maven.

https://github.com/jenkinsci/unleash-plugin has been archived.  That
is the Jenkins unleash Maven plugin. The last release of the Jenkins
unleash maven plugin was 3 years ago. Since the Jenkins plugin
repository has been archived, we can expect no new releases of the
plugin.

https://issues.jenkins.io/browse/JENKINS-70326 notes that the Jenkins
unleash maven plugin does not work with Java 17.

Let's deprecate the Jenkins plugin so that it is clear to users that
the Jenkins plugin is not being maintained and that the Apache Maven
plugin is not being maintained.

https://www.jenkins.io/doc/developer/plugin-governance/deprecating-or-removing-plugin/
says that the preferred way to deprecate a plugin is to add the
"deprecated" topic to the repository and to explain in the plugin
documentation why it is being deprecated.  That would mean allowing
writes to the repository again, just long enough to set the topic.

I assume that is more complicated than the alternate method, a pull
request to the update center repository to mark the repository as
deprecated.

https://github.com/jenkins-infra/plugins-wiki-docs/tree/master/unleash
is the documentation for the plugin.  I'll submit a pull request to
the wiki repository to note that the plugin is deprecated.