Let markewaite adopt slave-setup plugin
Plugin has an implied depenendency on WMI Windows Agents plugin as
noted at https://plugins.jenkins.io/slave-setup/#dependencies

The implied dependency can be resolved by delivering a new release of
the plugin that requires a more modern Jenkins version.

https://issues.jenkins.io/browse/JENKINS-70301 provides more details
on implied dependencies and why a new release is an easy way to
resolve the issue.

https://github.com/jenkinsci/slave-setup-plugin/pull/16 is the pull
request that modernizes the plugin, changes its display name to not
use the deprecated "slave" term, and prepares it to release a new
version that requires Jenkins 2.346.3 or newer.

https://www.jenkins.io/doc/developer/plugin-development/choosing-jenkins-baseline/
notes that 2.346.3 is a good choice for a minimum Jenkins version.

@peppelan and @wolfs are listed as current maintainers of the plugin.