Revert "Let markewaite adopt slave-setup plugin (#3053)"
https://github.com/jenkinsci/slave-setup-plugin/pull/21 notes why I'm
not ready to do any further work on the plugin.

Jenkins 2.386 will include a more general fix for the issue that I was
trying to resolve here, without risk of breaking the existing users of
the `slave-setup` plugin.

https://github.com/jenkinsci/jenkins/pull/7568 is the more general fix.

https://issues.jenkins.io/browse/JENKINS-70301 describes the implied
dependency that the slave-setup plugin has on the WMI Windows Agents
plugin.

This reverts commit 4d30c5e334decd2efb94597bb4d57df3c2ecf51a.