Let markewaite adopt copyartifact plugin
Need to release a new version that removes the test dependency on
compress-artifacts so that the copyartifact plugin can be updated in
the Jenkins plugin bill of materials.

https://github.com/jenkinsci/copyartifact-plugin/pull/163 removes the
test dependency.

https://github.com/jenkinsci/bom/pull/1689#issuecomment-1386975882 notes
how that removal will help unblock the update of bouncycastle api plugin
in the Jenkins plugin BOM.

https://github.com/jenkinsci/copyartifact-plugin/pull/164 is my pull
request to require Java 11 and Jenknis 2.361.4 or newer.  That will
resolve two additional pull requests:

* https://github.com/jenkinsci/copyartifact-plugin/pull/159
* https://github.com/jenkinsci/copyartifact-plugin/pull/162

https://github.com/jenkinsci/copyartifact-plugin shows the plugin is up
for adoption, so no additional approvals are required.