Let markewaite adopt Apache httpcomponents 4 API plugin
JGit 6.5 bundles newer versions of the httpcomponents libraries.  Rather than
use exclusions and other workarounds and then hope that there are not issues
when JGit tries to use the older libraries included in our current release of
the plugin, let me adopt this plugin, merge multiple pull requests, and release
a new version.

Pull requests that I intend to merge include:

* https://github.com/jenkinsci/apache-httpcomponents-client-4-api-plugin/pull/76
* https://github.com/jenkinsci/apache-httpcomponents-client-4-api-plugin/pull/109
* https://github.com/jenkinsci/apache-httpcomponents-client-4-api-plugin/pull/113
* https://github.com/jenkinsci/apache-httpcomponents-client-4-api-plugin/pull/118

Those pull requests are included in my prototype pull request along with
additional changes to match the current releases of the upstream library:

* https://github.com/jenkinsci/apache-httpcomponents-client-4-api-plugin/pull/120

Pull requests that will be closed because they are superseded by others:

* https://github.com/jenkinsci/apache-httpcomponents-client-4-api-plugin/pull/107
* https://github.com/jenkinsci/apache-httpcomponents-client-4-api-plugin/pull/112
* https://github.com/jenkinsci/apache-httpcomponents-client-4-api-plugin/pull/116
* https://github.com/jenkinsci/apache-httpcomponents-client-4-api-plugin/pull/119