Add markewaite as a (temporary) testcomplete plugin maintainer
@rinconedgar and @igor_filin are the maintainers
from SmartBear.  @rinconedgar has approved pull request
https://github.com/jenkinsci/testcomplete-plugin/pull/12 that
modernizes the plugin and resolves an already disclosed vulnerability
https://www.jenkins.io/security/advisory/2023-01-24/#SECURITY-2741 .
They would like to merge and release a new version of the plugin with
that fix, but @rinconedgar does not have merge permission to the GitHub
repository.

Allowing me as a temporary maintainer may also grant merge permission
to @rinconedgar.  If it does not, then we'll need someone to investigate
why the people listed in the repository permissions updater do not have
merge permission to the repository.

https://github.com/jenkinsci/testcomplete-plugin/pull/12 is the pull
request that I plan to merge.