Add markewaite to workflow-remote-loader-plugin
The most recent release of workflow-remote-loader-plugin packages
outdated versions of several jar files, including:

* groovy-all 2.0.5
* JGit 3.7.1
* Apache httpclient 4.4.1
* jsch 0.1.50

https://issues.jenkins.io/browse/JENKINS-62767 notes that JGit should
not be bundled by other plugins since it is provided by the git client
plugin as part of the git client plugin API.  JGit 3.7.1 bundled
inside this plugin does not seem to be harmful, but there is no
compelling reason to ship those libraries inside this plugin.

https://github.com/jenkinsci/workflow-remote-loader-plugin/pull/18 is
the pull request that I would like to merge and release.  I've
confirmed that the pull request generates an `hpi` file without those
extra dependencies.  The pull request also updates the plugin to
require Java 11 and use the plugin bill of materials.

The plugin is installed on over 3000 Jenkins controllers.

@escoem or @oleg-nenashev approval needed to add me as a plugin maintainer