Add markewaite as a maintainer of the adoptopenjdk plugin
I use this plugin as a fallback mechanism for those cases where I've not
installed a JDK already on my agents.

I'd like to merge the pull request:

* Test with Java 21
  https://github.com/jenkinsci/adoptopenjdk-plugin/pull/45

No release is needed, just a merge of the pull request.

Once that pull request is merged, I'll close the pull requests:

* Bump git changelist maven extension to 1.6
  https://github.com/jenkinsci/adoptopenjdk-plugin/pull/44
* Bump parent pom to 4.75
  https://github.com/jenkinsci/adoptopenjdk-plugin/pull/43
* Bump wiremock to 2.35.0
  https://github.com/jenkinsci/adoptopenjdk-plugin/pull/36

@hrmohr I'd love to have your approval to be added as a maintainer.