Add markewaite to ivy developers list
I'm not a user of the plugin but I want to adopt the plugin so that I
can merge https://github.com/jenkinsci/ivy-plugin/pull/87 and include
Java 21 testing in the regular development of the plugin.

With no current maintainers, that pull request won't be merged.
I'd rather merge it and reduce the number of open pull requests.

That pull request replaces multiple other pull requests, including:

* https://github.com/jenkinsci/ivy-plugin/pull/86
* https://github.com/jenkinsci/ivy-plugin/pull/85
* https://github.com/jenkinsci/ivy-plugin/pull/83
* https://github.com/jenkinsci/ivy-plugin/pull/80