Enable continuous delivery for parameterized trigger plugin
https://github.com/jenkinsci/parameterized-trigger-plugin/pull/366 is
the pull request to enable continuous delivery in the parameterized
trigger plugin.

I performed the previous release manually and would rather have the
simplification of automated releases since we now have a pending bug
fix pull request and a pending enhancement.

Bug fix:

* https://github.com/jenkinsci/parameterized-trigger-plugin/pull/357

Enhancement:

* https://github.com/jenkinsci/parameterized-trigger-plugin/pull/363

I'm not a user of the plugin, so I want to simplfy my contributions to
the plugin so that they will be released automatically.