Add markewaite as a gitee plugin maintainer
https://github.com/jenkinsci/gitee-plugin/pull/4 is the pull request
that I want to merge and release.

The Jackson 2 API plugin 2.16.1 has removed two constants that were
deprecated long ago. The removal of those constants will break
compilation of the gitee plugin and are expected to break runtime of
the gitee plugin as well.

The targeted pull request upgrades from Jackson 2 API 2.15.2 to Jackson
2 API 2.16.1 so that the API deletion of the constants is visible
at compile time.  A Jenkins controller with the Jackson 2 API plugin
2.16.1-373.ve709c6871598 installed may already make this change visible to
a gitee plugin at run time without any change to the plugin source code.

https://github.com/FasterXML/jackson-databind/pull/4162 removed the
constants from the Jackson 2 API after a long period of deprecation.
The changes were released in Jackson databind 2.16.0 and first
included in the matching Jenkins API plugin 2.16.1-373.ve709c6871598.

A similar change was needed in the GitLab plugin and is implemented in
https://github.com/jenkinsci/gitlab-plugin/pull/1606.

If @yashin-luo merges that pull request and releases the gitee plugin,
then I don't need to be added as a maintainer.  I've started this pull
request just in case @yashin-luo is not available and I need to wait
two weeks for plugin adoption.

https://issues.jenkins.io/browse/JENKINS-72482 is the bug report.