Add markewaite as a description setter maintainer
I'm a user of this plugin and there is only one listed maintainer
currently.

https://github.com/jenkinsci/description-setter-plugin/pull/26 is the
pull request that I plan to merge.  I don't plan to make major changes
to the plugin, only merge that pull request and assure that the
backlog of maintenance pull requests (like dependabot) are handled at
least once a month.

I'll probably modernize the plugin to move the documentation to the
GitHub repository and take other steps from the "Improve a plugin"
tutorial.  No commitment, but because I use the plugin, I will tend to
improve the plugin so that the maintenance is simpler.