Adopt Pegdown formatter plugin in order to deprecate it
The Pegdown formatter plugin provides markdown formatting
for Jenkins job descriptions and other text fields.
However, it was last released 12 years ago, has no
active maintainer, and has an open security vulnerability
https://jenkins.io/security/advisory/2019-08-07/#SECURITY-142

The markdown formatter plugin provides markdown formatting for Jenkins
job descriptions and other text fields.  It is a reasonable replacement
for the Pegdown formatter plugin, though without the many configuration
options available from the Pegdown formatter plugin.  The markdown
formatter plugin is actively maintained and used by Mark Waite.  It has
over 800 installations and the number of instalations is growing.

Let's deprecate the Pegdown formatter and include a link to the markdown
formatter in the deprecation so that users can continue using markdown
formatting and have a plugin without a security vulnerability.

Add @MarkEWaite as a maintainer so that he can merge:

* https://github.com/jenkinsci/pegdown-formatter-plugin/pull/2

Once I have merge permission, I'll apply the "deprecated" label to the
Pegdown formatter plugin repository and request that it be archived.

https://www.jenkins.io/doc/developer/plugin-governance/deprecating-or-removing-plugin/
is the instructions that I'm following for the deprecation process.