Add markewaite as a badge plugin maintainer
I was removed today from the GitHub developer team for this plugin.
That removal is correct because I am not listed in the repository
permissions updater for the plugin.

I had removed myself as a maintainer in November 2022
https://github.com/jenkins-infra/repository-permissions-updater/commit/eb71cad593051f9dea2cfa9eaafe8f66e143a44a
after @gounthar, @jmMeessen, and I had completed a Hacktoberfest
initiative to encourage new contributors and to recruit new
maintainers.

I've had enough fun with the most recent several pull requests that I
think I would enjoy returning as a maintainer of this plugin.

@strangelookingnerd, @mPokornyETM, and @bakito Would you accept me as
a maintainer?

It is not a problem if you would rather not have me as a maintainer in
the repository.  I understand that as well.  Too many maintainers can
be disruptive.