Add markewaite as ignore-commit-strategy maintainer
https://github.com/jenkinsci/ignore-committer-strategy-plugin/pull/10 is
a pull request to update the plugin to build with Java 11, 17, and 21.
The update adapts the plugin to changes in the branch API that happened
in 2019.

The pull request includes a few other pull requests:

* https://github.com/jenkinsci/ignore-committer-strategy-plugin/pull/3
* https://github.com/jenkinsci/ignore-committer-strategy-plugin/pull/5
* https://github.com/jenkinsci/ignore-committer-strategy-plugin/pull/7

I don't expect to make significant additional changes, but since I use
the plugin and it currently has an implied dependency on a plugin that
I would like to remove, I'm willing to adopt it.