Add markewaite as a last-changes plugin maintainer
https://github.com/jenkinsci/last-changes-plugin/pull/97 needs to be
merged and released so that users of the last-changes plugin are not
disrupted when they upgrade to git client 5.0.0 or git client plugin
6.0.0.

$The pull request was submitted in September.  There has been no response
from @rmpestano.
https://github.com/jenkinsci/last-changes-plugin/pull/97#issuecomment-2440212352

That pull request is a minimal change that adapts to git client plugin
5.0.0 and later without unnecessary changes.  Other changes to modernize
the plugin can be considered once the immediate issue with newer versions
of git client plugin has been resolved.

https://github.com/jenkinsci/last-changes-plugin/pull/98 includes several
steps to modernize the plugin, though it has not been tested yet.