Replace tasklist with checklist in PR template
GitHub announced that they are retiring
task lists as of 30 Apr 2025 in a blog post
https://github.blog/changelog/2025-02-18-github-issues-projects-february-18th-update/#tasklist-blocks-will-be-retired-and-replaced-with-sub-issues

They recommend a transition to sub-issues.  That can be done in a separate step.  Let's switch from tasklist to a simple checklist like we use for the Jenkins LTS release checklist so that the pull request description does not include the deprecation warning for task lists.

Testing done

Confirmed that simple checklists are still supported by GitHub and work well.