Drop markewaite as a last-changes maintainer
Plugin scans the SCM data for all builds of a job. That can cause slow
requests and performance issues.  I'm not willing to work on that
issue, so I'm placing the plugin for adoption.

Recent changes have modernized the plugin to Jakarta EE 9 and Java 17.
If someone else wants to adopt it, their adoption will start with a
good working base.