Drop markewaite as an authorize project plugin maintainer
I do not use the plugin and am not able to maintain it effectively.
It has been modernized and kept current with recent changes to Jenkins
core.  It is up for adoption and ready for others to adopt it.

https://old.stats.jenkins.io/pluginversions/authorize-project.html shows
that roughly 30$ of the installations are running the most recent release
on Jenkins core versions released in the last 9 months.