Add markewaite as a maintainer of build-with-parameters plugin
Plugin seems to be abandoned since there has not been a merge of the
EE 8 to EE 9 migration pull request:

* https://github.com/jenkinsci/build-with-parameters-plugin/pull/30

I will merge that and also merge my pull request for Java 25:

* https://github.com/jenkinsci/build-with-parameters-plugin/pull/32

Once those two pull requests are merged, I'll release a new version so
that the EE 8 to EE 9 migration changes are available for users.

@sugonyak is the current maintainer.  Would be nice if he can approve
me being added as a maintainer so that we don't need to wait the two
weeks for the timeout.