Add markewaite as a trilead maintainer
The trilead-ssh2 fork that is maintained by the Jenkins project has a bug.
That bug caused the regression documented in issue:

* https://github.com/jenkinsci/jenkins/issues/11312

The issue is fixed by a pull request to the trilead-ssh2 forked repository:

* https://github.com/jenkinsci/trilead-ssh2/pull/273

I need maintainer access to merge that pull request so that a new release
of the library can be created.  I've tested the following additonal pull
requests that should be merged as well:

* https://github.com/jenkinsci/trilead-ssh2/pull/268
* https://github.com/jenkinsci/trilead-ssh2/pull/269
* https://github.com/jenkinsci/trilead-ssh2/pull/270
* https://github.com/jenkinsci/trilead-ssh2/pull/271
* https://github.com/jenkinsci/trilead-ssh2/pull/272

Once the library is released, then I need permission to update the
trilead-api plugin to use the new library.  The repository for the
trilead-api plugin is:

* https://github.com/jenkinsci/trilead-api-plugin

Current committers to both repositories include:

* @kuisathaverat
* @mc1arke