Add markewaite as an audit-trail-plugin maintainer
The pull request to compile and test with Java 25 was opened two weeks
ago as:

* https://github.com/jenkinsci/audit-trail-plugin/pull/189

A reminder was posted a week ago asking the maintainers to review and
merge the request.

I don't plan to do significant work on the plugin, just enough to
merge that pull request and possibly two others:

* https://github.com/jenkinsci/audit-trail-plugin/pull/185
* https://github.com/jenkinsci/audit-trail-plugin/pull/188

I'm willing to be a maintainer of the plugin in order to assure that
it compiles and tests with Java 25.

Current maintainers are:

* @@cyrille-leclerc
* @PierreBtz
* @tomaszsek