Add markewaite as an ssh-steps-plugin maintainer
The pull request to compile and test with Java 25 was opened two weeks
ago as:

* https://github.com/jenkinsci/ssh-steps-plugin/pull/172

A reminder was posted asking the maintainers to review and merge the
request.

I don't plan to do significant work on the plugin, just enough to
merge that pull request and possibly:

* https://github.com/jenkinsci/ssh-steps-plugin/pull/171

I'm willing to be a maintainer of the plugin in order to assure that
it compiles and tests with Java 25.

Current maintainers are:

* @nrayapati